Create shape_predictor_68_face_landmarks.dat
Facial Landmark Detection: This file contains a pre-trained model used by the dlib library to detect 68 specific points (landmarks) on a human face. These points correspond to features like the corners of the eyes, the tip of the nose, the edges of the lips, and the outline of the face.